requirements.txt
Flask
Flask-APScheduler
praw
vaderSentiment
python-dotenv
openai
croniter
pyyaml
chromadb
sentence-transformers
requests
